requirements.txt
deepgram-sdk==2.12.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
flask==2.3.3
